fix: moved error enums to top-level `errors.rs` according to convention
