+++ block-producer/src/block_builder/mod.rs
@@ -8,17 +8,13 @@ use tracing::{debug, info, instrument};
 use crate::{
     batch_builder::batch::TransactionBatch,
     block::Block,
+    errors::BuildBlockError,
     store::{ApplyBlock, Store},
     COMPONENT, MAX_NUM_CREATED_NOTES_PER_BATCH,
 };
 
-pub mod errors;
-
 pub(crate) mod prover;
-use self::{
-    errors::BuildBlockError,
-    prover::{block_witness::BlockWitness, BlockProver},
-};
+use self::prover::{block_witness::BlockWitness, BlockProver};
 
 #[cfg(test)]
 mod tests;
